@@ -37,7 +37,7 @@
     "postcss": "^8.4.31",
     "prettier": "^3.0.3",
     "prettier-plugin-svelte": "^3.0.3",
-    "svelte": "^3.59.2",
+    "svelte": "^4.2.1",
     "svelte-check": "^3.5.2",
     "svelte-preprocess": "^5.0.4",
     "tailwindcss": "3.3.3",
