@@ -26,168 +26,147 @@
  * ```
  */
 declare module '$env/static/private' {
-	export const SHELL: string;
-	export const npm_command: string;
-	export const WINDOWID: string;
-	export const COLORTERM: string;
-	export const npm_package_scripts_generate_types: string;
-	export const XDG_CONFIG_DIRS: string;
-	export const npm_package_devDependencies__types_node: string;
-	export const XPC_FLAGS: string;
-	export const npm_package_repository_url: string;
-	export const fish_color_cancel: string;
-	export const npm_package_devDependencies_marked: string;
-	export const npm_package_exports___hooks_import: string;
-	export const fish_pager_color_completion: string;
-	export const npm_config_resolution_mode: string;
-	export const NODE: string;
-	export const fish_color_command: string;
-	export const npm_package_scripts_test_cross_platform_build: string;
-	export const __CFBundleIdentifier: string;
-	export const npm_package_exports___node_polyfills_types: string;
-	export const SSH_AUTH_SOCK: string;
-	export const npm_package_dependencies_mime: string;
-	export const npm_package_devDependencies_dts_buddy: string;
-	export const fish_color_valid_path: string;
-	export const __ABBR_TIPS_VALUES: string;
-	export const npm_package_peerDependencies_svelte: string;
-	export const HOMEBREW_PREFIX: string;
-	export const KITTY_PID: string;
-	export const npm_package_exports___import: string;
-	export const EDITOR: string;
-	export const npm_package_scripts_test_unit: string;
-	export const npm_package_dependencies_sade: string;
-	export const PWD: string;
-	export const NIX_PROFILES: string;
-	export const fish_color_end: string;
-	export const npm_package_devDependencies_vite: string;
-	export const LOGNAME: string;
+	export const npm_package_devDependencies__tailwindcss_typography: string;
 	export const fish_color_param: string;
-	export const MANPATH: string;
-	export const NIX_PATH: string;
-	export const __NIX_DARWIN_SET_ENVIRONMENT_DONE: string;
-	export const npm_package_exports___node_import: string;
-	export const KITTY_PUBLIC_KEY: string;
-	export const npm_package_dependencies_cookie: string;
-	export const npm_package_dependencies_sirv: string;
-	export const npm_package_dependencies_tiny_glob: string;
-	export const npm_package_bugs_url: string;
-	export const ABBR_TIPS_PROMPT: string;
-	export const npm_package_exports___node_types: string;
-	export const npm_package_exports___vite_types: string;
-	export const fish_pager_color_progress: string;
-	export const fish_color_operator: string;
-	export const COMMAND_MODE: string;
-	export const __ABBR_TIPS_KEYS: string;
-	export const HOMEBREW_BUNDLE_NO_LOCK: string;
-	export const HOME: string;
-	export const fish_color_cwd_root: string;
-	export const LANG: string;
-	export const npm_package_devDependencies_typescript: string;
-	export const npm_package_version: string;
 	export const STARSHIP_SHELL: string;
-	export const fish_color_host: string;
-	export const NIX_SSL_CERT_FILE: string;
-	export const KITTY_WINDOW_ID: string;
-	export const npm_package_files_0: string;
-	export const npm_package_scripts_check_all: string;
-	export const npm_package_files_1: string;
-	export const npm_package_files_2: string;
-	export const npm_package_files_3: string;
-	export const npm_package_repository_type: string;
-	export const npm_package_files_4: string;
-	export const npm_package_files_5: string;
-	export const npm_package_files_6: string;
-	export const npm_package_scripts_test_integration: string;
-	export const fish_color_redirection: string;
-	export const TMPDIR: string;
-	export const npm_package_exports___vite_import: string;
-	export const fish_color_history_current: string;
-	export const npm_package_devDependencies__types_set_cookie_parser: string;
-	export const npm_package_dependencies_kleur: string;
-	export const npm_package_devDependencies_rollup: string;
-	export const INIT_CWD: string;
-	export const fish_color_comment: string;
-	export const npm_package_scripts_test_cross_platform_dev: string;
-	export const fish_color_cwd: string;
-	export const npm_package_scripts_format: string;
-	export const STARSHIP_SESSION_KEY: string;
-	export const NIX_USER_PROFILE_DIR: string;
-	export const INFOPATH: string;
-	export const npm_lifecycle_script: string;
-	export const npm_package_description: string;
-	export const fish_color_selection: string;
-	export const npm_package_peerDependencies_vite: string;
-	export const TERM: string;
-	export const TERMINFO: string;
-	export const npm_package_name: string;
-	export const fish_color_autosuggestion: string;
-	export const npm_config_prefix: string;
-	export const npm_package_type: string;
-	export const USER: string;
-	export const npm_config_frozen_lockfile: string;
-	export const npm_package_devDependencies_vitest: string;
-	export const npm_package_exports___types: string;
-	export const fish_color_quote: string;
-	export const fish_color_normal: string;
-	export const npm_package_dependencies_undici: string;
-	export const npm_package_exports___package_json: string;
-	export const npm_package_homepage: string;
-	export const HOMEBREW_CELLAR: string;
-	export const npm_package_devDependencies__types_mime: string;
-	export const fish_color_user: string;
-	export const npm_package_dependencies_set_cookie_parser: string;
-	export const npm_package_dependencies__types_cookie: string;
-	export const npm_package_exports___node_polyfills_import: string;
-	export const npm_lifecycle_event: string;
-	export const SHLVL: string;
-	export const npm_package_dependencies_magic_string: string;
-	export const PAGER: string;
-	export const HOMEBREW_REPOSITORY: string;
-	export const npm_package_devDependencies__types_connect: string;
-	export const npm_package_devDependencies__types_sade: string;
-	export const XPC_SERVICE_NAME: string;
-	export const npm_config_user_agent: string;
-	export const npm_package_scripts_lint: string;
-	export const TERMINFO_DIRS: string;
-	export const PNPM_SCRIPT_SRC_DIR: string;
-	export const npm_execpath: string;
-	export const npm_package_devDependencies_svelte: string;
-	export const npm_package_scripts_test: string;
-	export const npm_package_devDependencies_svelte_preprocess: string;
-	export const fish_color_error: string;
-	export const KITTY_LISTEN_ON: string;
-	export const npm_package_scripts_generate_version: string;
-	export const npm_package_dependencies_esm_env: string;
-	export const npm_package_dependencies__sveltejs_vite_plugin_svelte: string;
-	export const npm_package_bin_svelte_kit: string;
-	export const XDG_DATA_DIRS: string;
-	export const npm_package_scripts_check: string;
-	export const PATH: string;
-	export const fish_color_match: string;
-	export const npm_package_dependencies_devalue: string;
-	export const npm_config_node_gyp: string;
-	export const __fish_nixos_env_preinit_sourced: string;
-	export const npm_package_exports___hooks_types: string;
-	export const npm_package_devDependencies__playwright_test: string;
-	export const npm_package_license: string;
-	export const npm_package_scripts_postinstall: string;
-	export const fish_color_search_match: string;
+	export const npm_package_devDependencies_prettier: string;
+	export const MANPATH: string;
+	export const NIX_PROFILES: string;
+	export const npm_package_dependencies_svelte_parallax: string;
 	export const base16_fish_shell_background: string;
-	export const npm_config_registry: string;
-	export const npm_package_repository_directory: string;
-	export const HOMEBREW_BUNDLE_FILE: string;
-	export const fish_pager_color_prefix: string;
-	export const ABBR_TIPS_REGEXES: string;
-	export const KITTY_INSTALLATION_DIR: string;
+	export const ABBR_TIPS_PROMPT: string;
+	export const npm_package_devDependencies_eslint_plugin_svelte: string;
+	export const fish_color_normal: string;
+	export const NODE: string;
+	export const INIT_CWD: string;
+	export const npm_package_devDependencies_prettier_plugin_svelte: string;
+	export const npm_package_devDependencies_typescript: string;
+	export const TERM: string;
+	export const npm_package_devDependencies_vite: string;
+	export const SHELL: string;
+	export const npm_package_devDependencies__types_cookie: string;
+	export const npm_package_devDependencies_vite_plugin_tailwind_purgecss: string;
+	export const HOMEBREW_REPOSITORY: string;
+	export const TMPDIR: string;
+	export const npm_package_scripts_lint: string;
+	export const npm_package_scripts_dev: string;
+	export const __ABBR_TIPS_KEYS: string;
+	export const WINDOWID: string;
 	export const fish_color_escape: string;
-	export const npm_node_execpath: string;
-	export const npm_config_engine_strict: string;
-	export const npm_package_types: string;
+	export const __ABBR_TIPS_VALUES: string;
+	export const npm_config_registry: string;
+	export const npm_package_devDependencies__sveltejs_kit: string;
+	export const __fish_nixos_env_preinit_sourced: string;
+	export const npm_package_devDependencies_svelte_preprocess: string;
+	export const npm_package_private: string;
+	export const PNPM_HOME: string;
+	export const USER: string;
+	export const fish_color_operator: string;
+	export const npm_package_license: string;
+	export const npm_package_scripts_check_watch: string;
+	export const COMMAND_MODE: string;
+	export const PNPM_SCRIPT_SRC_DIR: string;
+	export const fish_color_autosuggestion: string;
+	export const SSH_AUTH_SOCK: string;
 	export const __CF_USER_TEXT_ENCODING: string;
+	export const npm_package_devDependencies_postcss: string;
+	export const fish_color_search_match: string;
+	export const npm_package_devDependencies_eslint: string;
+	export const npm_execpath: string;
+	export const npm_package_devDependencies__typescript_eslint_eslint_plugin: string;
+	export const npm_package_devDependencies_web_vitals: string;
+	export const npm_package_devDependencies_tslib: string;
+	export const npm_package_devDependencies_svelte: string;
+	export const PAGER: string;
+	export const npm_package_devDependencies_emailjs: string;
+	export const npm_config_frozen_lockfile: string;
+	export const XDG_CONFIG_DIRS: string;
+	export const fish_color_match: string;
+	export const fish_pager_color_progress: string;
+	export const npm_package_devDependencies__typescript_eslint_parser: string;
+	export const PATH: string;
+	export const TERMINFO_DIRS: string;
+	export const fish_color_selection: string;
+	export const npm_config_engine_strict: string;
+	export const fish_color_user: string;
+	export const __CFBundleIdentifier: string;
+	export const npm_package_devDependencies__skeletonlabs_tw_plugin: string;
+	export const NIX_PATH: string;
+	export const PWD: string;
+	export const npm_package_devDependencies_tailwindcss: string;
+	export const npm_command: string;
+	export const npm_package_scripts_preview: string;
+	export const npm_lifecycle_event: string;
+	export const EDITOR: string;
+	export const KITTY_PID: string;
+	export const npm_package_devDependencies__skeletonlabs_skeleton: string;
+	export const npm_package_name: string;
+	export const LANG: string;
+	export const fish_pager_color_prefix: string;
+	export const fish_color_history_current: string;
+	export const npm_config_resolution_mode: string;
+	export const NODE_PATH: string;
+	export const npm_package_scripts_build: string;
+	export const fish_color_quote: string;
+	export const npm_package_devDependencies_vitest: string;
+	export const fish_color_cancel: string;
+	export const XPC_FLAGS: string;
+	export const npm_package_devDependencies__tailwindcss_forms: string;
+	export const NIX_SSL_CERT_FILE: string;
+	export const fish_color_error: string;
+	export const npm_config_node_gyp: string;
+	export const npm_package_devDependencies_eslint_config_prettier: string;
+	export const XPC_SERVICE_NAME: string;
+	export const npm_package_devDependencies__vercel_analytics: string;
+	export const npm_package_version: string;
+	export const npm_package_devDependencies_svelte_check: string;
+	export const fish_color_comment: string;
+	export const fish_color_cwd_root: string;
+	export const HOMEBREW_BUNDLE_NO_LOCK: string;
+	export const npm_package_devDependencies_autoprefixer: string;
+	export const SHLVL: string;
+	export const npm_package_type: string;
+	export const HOME: string;
+	export const ABBR_TIPS_REGEXES: string;
+	export const npm_package_scripts_test: string;
+	export const npm_package_devDependencies__types_animejs: string;
+	export const TERMINFO: string;
+	export const npm_package_devDependencies__sveltejs_adapter_vercel: string;
+	export const fish_color_end: string;
+	export const HOMEBREW_PREFIX: string;
+	export const fish_color_command: string;
+	export const fish_color_cwd: string;
+	export const STARSHIP_SESSION_KEY: string;
+	export const npm_package_scripts_format: string;
+	export const LOGNAME: string;
+	export const npm_lifecycle_script: string;
 	export const fish_pager_color_description: string;
-	export const npm_package_engines_node: string;
-	export const _: string;
+	export const XDG_DATA_DIRS: string;
+	export const fish_pager_color_completion: string;
+	export const KITTY_INSTALLATION_DIR: string;
+	export const npm_config_user_agent: string;
+	export const npm_package_devDependencies__iconify_svelte: string;
+	export const KITTY_WINDOW_ID: string;
+	export const npm_package_devDependencies__playwright_test: string;
+	export const HOMEBREW_CELLAR: string;
+	export const INFOPATH: string;
+	export const npm_package_devDependencies__types_node: string;
+	export const npm_package_dependencies_animejs: string;
+	export const npm_package_dependencies__floating_ui_dom: string;
+	export const KITTY_LISTEN_ON: string;
+	export const HOMEBREW_BUNDLE_FILE: string;
+	export const fish_color_redirection: string;
+	export const fish_color_valid_path: string;
+	export const npm_package_scripts_check: string;
+	export const npm_node_execpath: string;
+	export const npm_config_prefix: string;
+	export const NIX_USER_PROFILE_DIR: string;
+	export const fish_color_host: string;
+	export const KITTY_PUBLIC_KEY: string;
+	export const __NIX_DARWIN_SET_ENVIRONMENT_DONE: string;
+	export const npm_package_scripts_test_unit: string;
+	export const COLORTERM: string;
+	export const NODE_ENV: string;
 }
 
 /**
@@ -217,168 +196,147 @@ declare module '$env/static/public' {
  */
 declare module '$env/dynamic/private' {
 	export const env: {
-		SHELL: string;
-		npm_command: string;
-		WINDOWID: string;
-		COLORTERM: string;
-		npm_package_scripts_generate_types: string;
-		XDG_CONFIG_DIRS: string;
-		npm_package_devDependencies__types_node: string;
-		XPC_FLAGS: string;
-		npm_package_repository_url: string;
-		fish_color_cancel: string;
-		npm_package_devDependencies_marked: string;
-		npm_package_exports___hooks_import: string;
-		fish_pager_color_completion: string;
-		npm_config_resolution_mode: string;
-		NODE: string;
-		fish_color_command: string;
-		npm_package_scripts_test_cross_platform_build: string;
-		__CFBundleIdentifier: string;
-		npm_package_exports___node_polyfills_types: string;
-		SSH_AUTH_SOCK: string;
-		npm_package_dependencies_mime: string;
-		npm_package_devDependencies_dts_buddy: string;
-		fish_color_valid_path: string;
-		__ABBR_TIPS_VALUES: string;
-		npm_package_peerDependencies_svelte: string;
-		HOMEBREW_PREFIX: string;
-		KITTY_PID: string;
-		npm_package_exports___import: string;
-		EDITOR: string;
-		npm_package_scripts_test_unit: string;
-		npm_package_dependencies_sade: string;
-		PWD: string;
-		NIX_PROFILES: string;
-		fish_color_end: string;
-		npm_package_devDependencies_vite: string;
-		LOGNAME: string;
+		npm_package_devDependencies__tailwindcss_typography: string;
 		fish_color_param: string;
-		MANPATH: string;
-		NIX_PATH: string;
-		__NIX_DARWIN_SET_ENVIRONMENT_DONE: string;
-		npm_package_exports___node_import: string;
-		KITTY_PUBLIC_KEY: string;
-		npm_package_dependencies_cookie: string;
-		npm_package_dependencies_sirv: string;
-		npm_package_dependencies_tiny_glob: string;
-		npm_package_bugs_url: string;
-		ABBR_TIPS_PROMPT: string;
-		npm_package_exports___node_types: string;
-		npm_package_exports___vite_types: string;
-		fish_pager_color_progress: string;
-		fish_color_operator: string;
-		COMMAND_MODE: string;
-		__ABBR_TIPS_KEYS: string;
-		HOMEBREW_BUNDLE_NO_LOCK: string;
-		HOME: string;
-		fish_color_cwd_root: string;
-		LANG: string;
-		npm_package_devDependencies_typescript: string;
-		npm_package_version: string;
 		STARSHIP_SHELL: string;
-		fish_color_host: string;
-		NIX_SSL_CERT_FILE: string;
-		KITTY_WINDOW_ID: string;
-		npm_package_files_0: string;
-		npm_package_scripts_check_all: string;
-		npm_package_files_1: string;
-		npm_package_files_2: string;
-		npm_package_files_3: string;
-		npm_package_repository_type: string;
-		npm_package_files_4: string;
-		npm_package_files_5: string;
-		npm_package_files_6: string;
-		npm_package_scripts_test_integration: string;
-		fish_color_redirection: string;
-		TMPDIR: string;
-		npm_package_exports___vite_import: string;
-		fish_color_history_current: string;
-		npm_package_devDependencies__types_set_cookie_parser: string;
-		npm_package_dependencies_kleur: string;
-		npm_package_devDependencies_rollup: string;
-		INIT_CWD: string;
-		fish_color_comment: string;
-		npm_package_scripts_test_cross_platform_dev: string;
-		fish_color_cwd: string;
-		npm_package_scripts_format: string;
-		STARSHIP_SESSION_KEY: string;
-		NIX_USER_PROFILE_DIR: string;
-		INFOPATH: string;
-		npm_lifecycle_script: string;
-		npm_package_description: string;
-		fish_color_selection: string;
-		npm_package_peerDependencies_vite: string;
-		TERM: string;
-		TERMINFO: string;
-		npm_package_name: string;
-		fish_color_autosuggestion: string;
-		npm_config_prefix: string;
-		npm_package_type: string;
-		USER: string;
-		npm_config_frozen_lockfile: string;
-		npm_package_devDependencies_vitest: string;
-		npm_package_exports___types: string;
-		fish_color_quote: string;
-		fish_color_normal: string;
-		npm_package_dependencies_undici: string;
-		npm_package_exports___package_json: string;
-		npm_package_homepage: string;
-		HOMEBREW_CELLAR: string;
-		npm_package_devDependencies__types_mime: string;
-		fish_color_user: string;
-		npm_package_dependencies_set_cookie_parser: string;
-		npm_package_dependencies__types_cookie: string;
-		npm_package_exports___node_polyfills_import: string;
-		npm_lifecycle_event: string;
-		SHLVL: string;
-		npm_package_dependencies_magic_string: string;
-		PAGER: string;
-		HOMEBREW_REPOSITORY: string;
-		npm_package_devDependencies__types_connect: string;
-		npm_package_devDependencies__types_sade: string;
-		XPC_SERVICE_NAME: string;
-		npm_config_user_agent: string;
-		npm_package_scripts_lint: string;
-		TERMINFO_DIRS: string;
-		PNPM_SCRIPT_SRC_DIR: string;
-		npm_execpath: string;
-		npm_package_devDependencies_svelte: string;
-		npm_package_scripts_test: string;
-		npm_package_devDependencies_svelte_preprocess: string;
-		fish_color_error: string;
-		KITTY_LISTEN_ON: string;
-		npm_package_scripts_generate_version: string;
-		npm_package_dependencies_esm_env: string;
-		npm_package_dependencies__sveltejs_vite_plugin_svelte: string;
-		npm_package_bin_svelte_kit: string;
-		XDG_DATA_DIRS: string;
-		npm_package_scripts_check: string;
-		PATH: string;
-		fish_color_match: string;
-		npm_package_dependencies_devalue: string;
-		npm_config_node_gyp: string;
-		__fish_nixos_env_preinit_sourced: string;
-		npm_package_exports___hooks_types: string;
-		npm_package_devDependencies__playwright_test: string;
-		npm_package_license: string;
-		npm_package_scripts_postinstall: string;
-		fish_color_search_match: string;
+		npm_package_devDependencies_prettier: string;
+		MANPATH: string;
+		NIX_PROFILES: string;
+		npm_package_dependencies_svelte_parallax: string;
 		base16_fish_shell_background: string;
-		npm_config_registry: string;
-		npm_package_repository_directory: string;
-		HOMEBREW_BUNDLE_FILE: string;
-		fish_pager_color_prefix: string;
-		ABBR_TIPS_REGEXES: string;
-		KITTY_INSTALLATION_DIR: string;
+		ABBR_TIPS_PROMPT: string;
+		npm_package_devDependencies_eslint_plugin_svelte: string;
+		fish_color_normal: string;
+		NODE: string;
+		INIT_CWD: string;
+		npm_package_devDependencies_prettier_plugin_svelte: string;
+		npm_package_devDependencies_typescript: string;
+		TERM: string;
+		npm_package_devDependencies_vite: string;
+		SHELL: string;
+		npm_package_devDependencies__types_cookie: string;
+		npm_package_devDependencies_vite_plugin_tailwind_purgecss: string;
+		HOMEBREW_REPOSITORY: string;
+		TMPDIR: string;
+		npm_package_scripts_lint: string;
+		npm_package_scripts_dev: string;
+		__ABBR_TIPS_KEYS: string;
+		WINDOWID: string;
 		fish_color_escape: string;
-		npm_node_execpath: string;
-		npm_config_engine_strict: string;
-		npm_package_types: string;
+		__ABBR_TIPS_VALUES: string;
+		npm_config_registry: string;
+		npm_package_devDependencies__sveltejs_kit: string;
+		__fish_nixos_env_preinit_sourced: string;
+		npm_package_devDependencies_svelte_preprocess: string;
+		npm_package_private: string;
+		PNPM_HOME: string;
+		USER: string;
+		fish_color_operator: string;
+		npm_package_license: string;
+		npm_package_scripts_check_watch: string;
+		COMMAND_MODE: string;
+		PNPM_SCRIPT_SRC_DIR: string;
+		fish_color_autosuggestion: string;
+		SSH_AUTH_SOCK: string;
 		__CF_USER_TEXT_ENCODING: string;
+		npm_package_devDependencies_postcss: string;
+		fish_color_search_match: string;
+		npm_package_devDependencies_eslint: string;
+		npm_execpath: string;
+		npm_package_devDependencies__typescript_eslint_eslint_plugin: string;
+		npm_package_devDependencies_web_vitals: string;
+		npm_package_devDependencies_tslib: string;
+		npm_package_devDependencies_svelte: string;
+		PAGER: string;
+		npm_package_devDependencies_emailjs: string;
+		npm_config_frozen_lockfile: string;
+		XDG_CONFIG_DIRS: string;
+		fish_color_match: string;
+		fish_pager_color_progress: string;
+		npm_package_devDependencies__typescript_eslint_parser: string;
+		PATH: string;
+		TERMINFO_DIRS: string;
+		fish_color_selection: string;
+		npm_config_engine_strict: string;
+		fish_color_user: string;
+		__CFBundleIdentifier: string;
+		npm_package_devDependencies__skeletonlabs_tw_plugin: string;
+		NIX_PATH: string;
+		PWD: string;
+		npm_package_devDependencies_tailwindcss: string;
+		npm_command: string;
+		npm_package_scripts_preview: string;
+		npm_lifecycle_event: string;
+		EDITOR: string;
+		KITTY_PID: string;
+		npm_package_devDependencies__skeletonlabs_skeleton: string;
+		npm_package_name: string;
+		LANG: string;
+		fish_pager_color_prefix: string;
+		fish_color_history_current: string;
+		npm_config_resolution_mode: string;
+		NODE_PATH: string;
+		npm_package_scripts_build: string;
+		fish_color_quote: string;
+		npm_package_devDependencies_vitest: string;
+		fish_color_cancel: string;
+		XPC_FLAGS: string;
+		npm_package_devDependencies__tailwindcss_forms: string;
+		NIX_SSL_CERT_FILE: string;
+		fish_color_error: string;
+		npm_config_node_gyp: string;
+		npm_package_devDependencies_eslint_config_prettier: string;
+		XPC_SERVICE_NAME: string;
+		npm_package_devDependencies__vercel_analytics: string;
+		npm_package_version: string;
+		npm_package_devDependencies_svelte_check: string;
+		fish_color_comment: string;
+		fish_color_cwd_root: string;
+		HOMEBREW_BUNDLE_NO_LOCK: string;
+		npm_package_devDependencies_autoprefixer: string;
+		SHLVL: string;
+		npm_package_type: string;
+		HOME: string;
+		ABBR_TIPS_REGEXES: string;
+		npm_package_scripts_test: string;
+		npm_package_devDependencies__types_animejs: string;
+		TERMINFO: string;
+		npm_package_devDependencies__sveltejs_adapter_vercel: string;
+		fish_color_end: string;
+		HOMEBREW_PREFIX: string;
+		fish_color_command: string;
+		fish_color_cwd: string;
+		STARSHIP_SESSION_KEY: string;
+		npm_package_scripts_format: string;
+		LOGNAME: string;
+		npm_lifecycle_script: string;
 		fish_pager_color_description: string;
-		npm_package_engines_node: string;
-		_: string;
+		XDG_DATA_DIRS: string;
+		fish_pager_color_completion: string;
+		KITTY_INSTALLATION_DIR: string;
+		npm_config_user_agent: string;
+		npm_package_devDependencies__iconify_svelte: string;
+		KITTY_WINDOW_ID: string;
+		npm_package_devDependencies__playwright_test: string;
+		HOMEBREW_CELLAR: string;
+		INFOPATH: string;
+		npm_package_devDependencies__types_node: string;
+		npm_package_dependencies_animejs: string;
+		npm_package_dependencies__floating_ui_dom: string;
+		KITTY_LISTEN_ON: string;
+		HOMEBREW_BUNDLE_FILE: string;
+		fish_color_redirection: string;
+		fish_color_valid_path: string;
+		npm_package_scripts_check: string;
+		npm_node_execpath: string;
+		npm_config_prefix: string;
+		NIX_USER_PROFILE_DIR: string;
+		fish_color_host: string;
+		KITTY_PUBLIC_KEY: string;
+		__NIX_DARWIN_SET_ENVIRONMENT_DONE: string;
+		npm_package_scripts_test_unit: string;
+		COLORTERM: string;
+		NODE_ENV: string;
 		[key: `PUBLIC_${string}`]: undefined;
 		[key: `${string}`]: string | undefined;
 	}
